requirements.txt
django
drf-spectacular
djangorestframework